refactor(BatchModule): move subcomponents into module
