+++ src/components/Datatable/BatchModule/BatchActions/BatchActions.tsx
@@ -3,21 +3,21 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { isNotUndefined } from 'ramda-adjunct';
 
-import { pxToRem } from '../../../utils/helpers';
-import { FlexContainer } from '../../FlexContainer';
-import { Button } from '../../Button';
-import { ButtonVariants } from '../../Button/Button.enums';
+import { ColorTypes } from '../../../../theme/colors.enums';
+import { pxToRem } from '../../../../utils/helpers';
+import { FlexContainer } from '../../../FlexContainer';
+import { Button } from '../../../Button';
+import { ButtonVariants } from '../../../Button/Button.enums';
+import { Icon } from '../../../Icon';
+import { SSCIconNames } from '../../../Icon/Icon.enums';
 import {
   AbsoluteLinkActionKind,
   ActionPropType,
   ActionWithSubactions,
   RelativeLinkActionKind,
-} from '../types/Action.types';
+} from '../../types/Action.types';
+import { Dropdown } from '../../Dropdown';
 import { BatchActionsProps } from './BatchActions.types';
-import { Dropdown } from '../Dropdown';
-import { Icon } from '../../Icon';
-import { SSCIconNames } from '../../Icon/Icon.enums';
-import { ColorTypes } from '../../../theme/colors.enums';
 
 const BatchActionButton = styled(Button).attrs((props) => ({
   variant: ButtonVariants.text,
