+++ src/components/Datatable/BatchModule/BatchActions/BatchActions.types.ts
@@ -1,4 +1,4 @@
-import { Action } from '../types/Action.types';
+import { Action } from '../../types/Action.types';
 
 export interface BatchActionsProps {
   actions: Action[];
