+++ src/components/Datatable/BatchModule/BatchModule.types.ts
@@ -1,5 +1,5 @@
-import { BatchActionsProps } from '../BatchActions/BatchActions.types';
-import { ElementCounterProps } from '../ElementCounter/ElementCounter.types';
+import { BatchActionsProps } from './BatchActions/BatchActions.types';
+import { ElementCounterProps } from './ElementCounter/ElementCounter.types';
 
 export interface BatchModuleProps
   extends BatchActionsProps,
